requirements.txt
langchain==0.3.12
langchain-community==0.3.12
langchain-core==0.3.25
langchain-text-splitters==0.3.3
langchain-huggingface==0.1.2
langchain-groq==0.2.1
langgraph==0.2.60
faiss-cpu
python-dotenv
streamlit
pypdf
sentence-transformers
faiss-cpu
python-dotenv
streamlit
pypdf
sentence-transformers
